feat: rewrite and rename cjsToEsm file
